Bug fix to vs issues
