--- Temple Course Helper/Temple Course Helper.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31919.166
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp1", "WindowsFormsApp1\WindowsFormsApp1.csproj", "{6F6CA410-2D64-4903-BEFB-789CCA4E319B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6F6CA410-2D64-4903-BEFB-789CCA4E319B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F6CA410-2D64-4903-BEFB-789CCA4E319B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F6CA410-2D64-4903-BEFB-789CCA4E319B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6F6CA410-2D64-4903-BEFB-789CCA4E319B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4D89410C-C2D6-4F9B-BBE8-32356442B649}
-	EndGlobalSection
-EndGlobal
