+++ Temple Course Helper/TempleCourseHelper.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31919.166
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TempleCourseHelper", "TempleCourseHelper\TempleCourseHelper.csproj", "{6FDA2CCF-2D41-4572-8030-BE365D1A837F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6FDA2CCF-2D41-4572-8030-BE365D1A837F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6FDA2CCF-2D41-4572-8030-BE365D1A837F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6FDA2CCF-2D41-4572-8030-BE365D1A837F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6FDA2CCF-2D41-4572-8030-BE365D1A837F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4A39096F-123D-4202-99FC-8E7C17DEF141}
+	EndGlobalSection
+EndGlobal
