+++ Temple Course Helper/TempleCourseHelper/CourseDetails.cs
