+++ Temple Course Helper/TempleCourseHelper/Properties/Resources.resx
