+++ Temple Course Helper/TempleCourseHelper/Properties/Settings.settings
