+++ Temple Course Helper/TempleCourseHelper/Worker.cs
