+++ Temple Course Helper/TempleCourseHelper/packages.config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.IdentityModel.Logging" version="1.1.2" targetFramework="net472" />
+  <package id="Microsoft.IdentityModel.Tokens" version="5.1.2" targetFramework="net472" />
+  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net472" />
+  <package id="Selenium.WebDriver" version="4.1.0" targetFramework="net472" />
+  <package id="System.IdentityModel.Tokens.Jwt" version="5.1.2" targetFramework="net472" />
+  <package id="Twilio" version="5.73.0" targetFramework="net472" />
+</packages>
